Custom Collision, Enemy attacks
Custom collision types added
HitBox - for the hitbox volume
AttackRange - For the volume in which the actor can attack other actors if they overlap this volume
Projectile - used for projectiles spawned by actors and can only overlap hitbox and WorldStatic

Enemies can now attack the player if the player is within their attack range and the player can also shoot projectiles in order to kill the enemy.
Currently no melee attack for player however that will be added soon.